Update VerticalFederalForm.cs
Deviations need to be recalculated before writing to file.  This was important for the case where filed data is loaded into the application.  Without a recalculation the extreme deviation  and variation would not be calculated correctly because the Adev,Bdev,Ddev,Edev are not stored in the raw data ouput file.